clean code, add comments
