+++ components/Logo/Logo.css.ts
@@ -1,12 +1,5 @@
 import { style } from "@vanilla-extract/css";
 
-export const outer = style({
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  width: "100%",
-});
-
 export const imageContainer = style({
   width: "50%",
   maxWidth: "600px",
