+++ components/index.css.ts
@@ -1,0 +1,8 @@
+import { style } from "@vanilla-extract/css";
+
+export const outer = style({
+  display: "flex",
+  flexDirection: "column",
+  alignItems: "center",
+  width: "100%",
+});
