+++ pages/index.tsx
@@ -1,8 +1,5 @@
 import { RawJoke, JokeParser } from "components/JokeParser";
-import NextImage from "next/image";
-import logo from "../public/assets/images/downtimeLogo.png";
-import * as styles from "../components/Logo/Logo.css";
-import { createClient } from "contentful";
+import * as styles from "components/index.css"
 import { generateFeed } from "../scripts/gen-rss";
 import getJokes from "lib/jokes";
 import Logo from "components/Logo";
@@ -13,9 +10,6 @@ export interface HomeProps {
 export default function Home({ jokes }: HomeProps) {
   return (
     <div className={styles.outer}>
-      {/* <div className={styles.imageContainer}>
-      <NextImage src={logo} alt="downtime logo"/>
-    </div> */}
       <Logo />
       <JokeParser jokes={jokes} />
     </div>
@@ -24,15 +18,6 @@ export default function Home({ jokes }: HomeProps) {
 
 export async function getStaticProps() {
   const jokes = await getJokes();
-  // const contentfulClient = createClient({
-  //   accessToken: `${process.env.CONTENTFUL_DELIVERY_ACCESS_TOKEN}`,
-  //   space: `${process.env.CONTENTFUL_SPACE_ID}`,
-  // });
-
-  // const res = await contentfulClient.getEntries({
-  //   content_type: "joke",
-  //   order: "-fields.pubDate",
-  // });
 
   await generateFeed(jokes as RawJoke[]);
